* Remove .NET Standard 2.1 targeting * Consolidate namespaces * General cleanup
