--- Salar.BinaryBuffers/Salar.BinaryBuffers.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;net6.0</TargetFrameworks>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <Description>A high performance implementation of BinaryReader and BinaryWriter which works on binary arrays directly by eliminating the need of a middle man Stream.</Description>
-    <Copyright>Copyright 2019</Copyright>
-    <PackageTags>Binary BinaryReader BinaryWriter</PackageTags>
-    <Authors>Salar Khalilzadeh</Authors>
-    <PackageProjectUrl>https://github.com/salarcode/BinaryBuffers</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/salarcode/BinaryBuffers</RepositoryUrl>
-    <RepositoryType></RepositoryType>
-    <AssemblyVersion>1.1.0.0</AssemblyVersion>
-    <FileVersion>1.1.0.0</FileVersion>
-    <Version>1.1.0</Version>
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <PackageReference Include="System.Memory" Version="4.5.4" />
-    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\LICENSE">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-</Project>
