+++ benchmarks/BinaryBuffers.Benchmark/BinaryBuffers.Benchmark.csproj
@@ -2,8 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>net48;net6.0</TargetFrameworks>
-    <Configurations>Release</Configurations>
+    <TargetFrameworks>net6.0</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
@@ -17,7 +16,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Salar.BinaryBuffers\Salar.BinaryBuffers.csproj" />
+    <ProjectReference Include="..\..\src\BinaryBuffers\BinaryBuffers.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Configurations\" />
   </ItemGroup>
 
 </Project>
