+++ tests/BinaryBuffers.Tests/BinaryBufferWriterTests.cs
@@ -1,45 +1,45 @@
-﻿using System;
-using System.IO;
-using System.Text;
-using Xunit;
-
-namespace Salar.BinaryBuffers.Tests
+﻿namespace BinaryBuffers.Tests
 {
-	public class BinaryBufferReaderTests : IDisposable
+    using Xunit;
+
+    using System;
+    using System.IO;
+    using System.Text;
+
+    public class BinaryBufferWriterTests : IDisposable
 	{
         private readonly MemoryStream _memoryStream;
-		private readonly BinaryWriter _writer;
-		private readonly BinaryBufferReader _bufferReader;
+		private readonly BinaryReader _reader;
+		private readonly BinaryBufferWriter _bufferWriter;
 
-		public BinaryBufferReaderTests()
+		public BinaryBufferWriterTests()
 		{
             var data = new byte[1024];
 			_memoryStream = new MemoryStream(data);
-			_writer = new BinaryWriter(_memoryStream, Encoding.UTF8, true);
-			_bufferReader = new BinaryBufferReader(data);
+			_reader = new BinaryReader(_memoryStream, Encoding.UTF8, true);
+			_bufferWriter = new BinaryBufferWriter(data);
 		}
 
 		private void Reset()
 		{
 			_memoryStream.Position = 0;
-			_memoryStream.SetLength(0);
-			_bufferReader.Position = 0;
+			_bufferWriter.Position = 0;
 		}
 
 		[Theory]
 		[InlineData(0)]
 		[InlineData(-1)]
 		[InlineData(byte.MaxValue)]
-		[InlineData(Int16.MinValue)]
-		[InlineData(Int16.MinValue / 2)]
-		[InlineData(Int16.MaxValue)]
-		[InlineData(Int16.MaxValue / 2)]
-		public void ReadInt16(Int16 input)
+		[InlineData(short.MinValue)]
+		[InlineData(short.MinValue / 2)]
+		[InlineData(short.MaxValue)]
+		[InlineData(short.MaxValue / 2)]
+		public void ReadInt16(short input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadInt16();
+			var val = _reader.ReadInt16();
 
 			Assert.Equal(input, val);
 		}
@@ -47,14 +47,14 @@ namespace Salar.BinaryBuffers.Tests
 		[Theory]
 		[InlineData(0)]
 		[InlineData(byte.MaxValue)]
-		[InlineData(UInt16.MaxValue)]
-		[InlineData(UInt16.MaxValue / 2)]
-		public void ReadUInt16(UInt16 input)
+		[InlineData(ushort.MaxValue)]
+		[InlineData(ushort.MaxValue / 2)]
+		public void ReadUInt16(ushort input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadUInt16();
+			var val = _reader.ReadUInt16();
 
 			Assert.Equal(input, val);
 		}
@@ -72,9 +72,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadInt32(int input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadInt32();
+			var val = _reader.ReadInt32();
 
 			Assert.Equal(input, val);
 		}
@@ -88,9 +88,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadUInt32(uint input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadUInt32();
+			var val = _reader.ReadUInt32();
 
 			Assert.Equal(input, val);
 		}
@@ -112,9 +112,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadInt64(long input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadInt64();
+			var val = _reader.ReadInt64();
 
 			Assert.Equal(input, val);
 		}
@@ -130,9 +130,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadUInt64(ulong input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadUInt64();
+			var val = _reader.ReadUInt64();
 
 			Assert.Equal(input, val);
 		}
@@ -156,9 +156,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadSingle(float input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadSingle();
+			var val = _reader.ReadSingle();
 
 			Assert.Equal(input, val);
 		}
@@ -182,9 +182,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadDouble(double input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadDouble();
+			var val = _reader.ReadDouble();
 
 			Assert.Equal(input, val);
 		}
@@ -205,9 +205,9 @@ namespace Salar.BinaryBuffers.Tests
 		{
 			Reset();
 			var input = Convert.ToDecimal(inputObj);
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadDecimal();
+			var val = _reader.ReadDecimal();
 
 			Assert.Equal(input, val);
 		}
@@ -220,9 +220,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadByte(byte input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadByte();
+			var val = _reader.ReadByte();
 
 			Assert.Equal(input, val);
 		}
@@ -238,37 +238,12 @@ namespace Salar.BinaryBuffers.Tests
 			var element = (byte) (input / 2);
 			Array.Fill(buff, element);
 
-			_writer.Write(buff);
+			_bufferWriter.Write(buff);
 
-			var val = _bufferReader.ReadBytes(input);
+			var val = _reader.ReadBytes(input);
 
 			Assert.Equal(input, val.Length);
 			Assert.Equal(buff, val);
-		}
-
-
-		[Theory]
-		[InlineData(0)]
-		[InlineData(1)]
-		[InlineData(150)]
-		public void ReadSpan(int input)
-		{
-			Reset();
-			var buff = new byte[input];
-			var element = (byte) (input / 2);
-			var elementSpan = new ReadOnlySpan<byte>(buff);
-			Array.Fill(buff, element);
-
-			_writer.Write(buff);
-
-			var val = _bufferReader.ReadSpan(input);
-
-			Assert.Equal(input, val.Length);
-
-			for (int i = 0; i < buff.Length; i++)
-			{
-				Assert.Equal(buff[i], val[i]);
-			}
 		}
 
 		[Theory]
@@ -282,9 +257,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadSByte(sbyte input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadSByte();
+			var val = _reader.ReadSByte();
 
 			Assert.Equal(input, val);
 		}
@@ -295,9 +270,9 @@ namespace Salar.BinaryBuffers.Tests
 		public void ReadBoolean(bool input)
 		{
 			Reset();
-			_writer.Write(input);
+			_bufferWriter.Write(input);
 
-			var val = _bufferReader.ReadBoolean();
+			var val = _reader.ReadBoolean();
 
 			Assert.Equal(input, val);
 		}
