+++ tests/BinaryBuffers.Tests/BinaryBuffers.Tests.csproj
@@ -5,7 +5,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
+      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
       <PackageReference Include="xunit" Version="2.4.1" />
       <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
         <PrivateAssets>all</PrivateAssets>
@@ -14,7 +14,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Salar.BinaryBuffers\Salar.BinaryBuffers.csproj" />
+      <ProjectReference Include="..\..\src\BinaryBuffers\BinaryBuffers.csproj" />
     </ItemGroup>
 
 </Project>
